skill assessment map is ready
A concept that I am stealing from Boreal Alyph. Basically, its set up so that you have to go in, turn a generator on to open the exit. Anyone who wants to do it can download the map, add their name to it and get to work. It's really just to show that you can design a fun an engaging map. You can redo any part of the map but the exit tunnel, though you can make the map look nicer if you want.